+++ blocks/browse-rail/browse-rail-utils.js
@@ -1,51 +1,3 @@
-// Utility function to check if the element is visible on the page.
-export function isVisible(element) {
-  const style = getComputedStyle(element);
-  return style.display !== 'none' && style.visibility !== 'hidden';
-}
-
-// Utility function to create dynamic list items
-export function createListItem(item) {
-  const li = document.createElement('li');
-  li.innerHTML = `<a href="${item.path}">${item.title}</a>`;
-  return li;
-}
-
-// Utility function to toggle visibility of items
-export function toggleItemVisibility(itemList, startIndex, show) {
-  // eslint-disable-next-line no-plusplus
-  for (let i = startIndex; i < itemList.length; i++) {
-    itemList[i].classList.toggle('hidden', !show);
-  }
-}
-
-// Utility function to set link visibility
-export function setLinkVisibility(linkId, show) {
-  const linkElement = document.getElementById(linkId);
-  if (linkElement) {
-    linkElement.style.display = show ? 'block' : 'none';
-  }
-}
-
-// Function to check if current page has sub-pages
-export function hasDirectLeafNodes(jsonData, currentPage) {
-  const directLeafNodes = jsonData.filter(
-    (item) =>
-      item.path.startsWith(currentPage) &&
-      item.path !== currentPage &&
-      !item.path.substring(currentPage.length + 1).includes('/'),
-  );
-
-  return directLeafNodes.length > 0;
-}
-
-// Utility Function to get the path until a specific level
-export function getPathUntilLevel(originalUrl, levels) {
-  const pathSegments = originalUrl.split('/');
-  const resultPath = pathSegments.slice(0, levels + 1).join('/');
-  return resultPath;
-}
-
 // Utility function to filter sub-pages under a given path
 export const filterSubPages = (data, basePath) =>
   data.filter((page) => page.path.startsWith(basePath) && page.path !== basePath);
